+++ packages/main/src/components/AnalyticalTable/defaults/Column/ColumnHeaderModal.tsx
@@ -1,3 +1,4 @@
+import IconMode from '@ui5/webcomponents/dist/types/IconMode.js';
 import ListItemType from '@ui5/webcomponents/dist/types/ListItemType.js';
 import PopoverHorizontalAlign from '@ui5/webcomponents/dist/types/PopoverHorizontalAlign.js';
 import PopoverPlacement from '@ui5/webcomponents/dist/types/PopoverPlacement.js';
@@ -218,7 +219,7 @@ export const ColumnHeaderModal = (instance: TableInstanceWithPopoverProps) => {
               <Icon
                 name={iconFilter}
                 className={classNames.filterIcon}
-                aria-hidden
+                mode={IconMode.Decorative}
                 style={{
                   minWidth: filterStyles.iconDimensions,
                   minHeight: filterStyles.iconDimensions