+++ packages/main/src/components/AnalyticalTable/hooks/useToggleRowExpand.ts
@@ -1,13 +1,20 @@
-import { enrichEventWithDetails } from '@ui5/webcomponents-react-base';
+import announce from '@ui5/webcomponents-base/dist/util/InvisibleMessage.js';
+import { debounce, enrichEventWithDetails } from '@ui5/webcomponents-react-base';
 import type { ReactTableHooks, RowType, TableInstance } from '../types/index.js';
+
+// debounce announce to prevent excessive successive announcements
+const debouncedAnnounce = debounce((announcement: string) => {
+  announce(announcement, 'Polite');
+}, 200);
 
 const getToggleRowExpandedProps = (
   rowProps,
   { row, instance, userProps }: { row: RowType; instance: TableInstance; userProps: Record<string, any> }
 ) => {
   const { manualGroupBy } = instance;
-  const { onRowExpandChange, isTreeTable, renderRowSubComponent, alwaysShowSubComponent } =
+  const { onRowExpandChange, isTreeTable, renderRowSubComponent, alwaysShowSubComponent, translatableTexts } =
     instance.webComponentsReactProperties;
+
   const onClick = (e, noPropagation = true) => {
     if (noPropagation) {
       e.stopPropagation();
@@ -30,6 +37,11 @@ const getToggleRowExpandedProps = (
       );
     }
     row.toggleRowExpanded();
+    // cannot use ROW_X_COLLAPSED/ROW_X_EXPANDED here,
+    // as retrieving the index of the row is not easily possible here and has performance implications
+    debouncedAnnounce(
+      !row.isExpanded ? translatableTexts.rowExpandedAnnouncementText : translatableTexts.rowCollapsedAnnouncementText
+    );
   };
   const onKeyDown = (e) => {
     if (e.code === 'F4') {